First working prototype! (it was so quick after all the Eclipse/Android dependencies hurdles).
This is a very cool first demo that put all the components together:
Nexus-7 secured on the tilt system mounted on the Roomba create and
connected to the IOIO with custom shield.
Skype running on the tablet for visual feedback

Issues:
- Service notification "stop" button does not appear anymore when there
are more notifications (hard to stop the service)
- The system seems to stop responding after a while (only 2 runs so far,
so no clue yet)
- Everything configuration wise (ip, ports, pin) is hard-coded
- PageUp/PageDown (used for tilt control) do not exist as real key on a
laptop keyboard
- DriveVelocityMsg second argument is used to transmit a radius while
the spec says we want curvature. The same applies for unit: basically
DriveVelocityMsg simply transport the numbers we want to give to the
current IoioRoombaCreate.
- Quite a bit of audio is captured by the microphone (so close to high
pitch motors)
- The whole thing is not resilient to restart or comm errors.